+++ src/lib.rs
@@ -51,6 +51,7 @@ pub use crate::algorithms::connectivity::biconnected::{
 };
 pub use crate::algorithms::connectivity::bridges::bridges;
 pub use crate::algorithms::connectivity::components::{ConnectedComponents, connected_components};
+pub use crate::algorithms::connectivity::decompose::decompose;
 pub use crate::algorithms::connectivity::is_biconnected::is_biconnected;
 pub use crate::algorithms::connectivity::reachability::count_reachable;
 pub use crate::algorithms::connectivity::reachability_matrix::reachability_matrix;